user guards set

+++ src/users/users.controller.ts
@@ -1,9 +1,10 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards, Request } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards, Request, BadRequestException } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { RoleGuard } from 'src/common/guards/roleGuard';
 import { Roles } from 'src/common/guards/roles.decorator';
+import { AuthGuard } from 'src/common/guards/authGuard';
 
 @Controller('users')
 export class UsersController {
@@ -15,7 +16,7 @@ export class UsersController {
   }
 
   @UseGuards(RoleGuard)
-  @Roles('user')
+  @Roles('admin')
   @Get()
   async findAll(@Request() req:any) {
     // const userData = req.user   //user data
@@ -24,18 +25,33 @@ export class UsersController {
     return await this.usersService.findAll();
   }
 
+  @UseGuards(AuthGuard)
   @Get(':id')
-  findOne(@Param('id') id: string) {
+  findOne(@Param('id') id: string,@Request() req:any) {
+    const authData = req.user
+    if(authData.role !== 'admin' && authData.id !== +id){
+      throw new BadRequestException("yu cant see others' profile")
+    }
     return this.usersService.findOne(+id);
   }
 
+  @UseGuards(AuthGuard)
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
+  update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto,@Request() req:any) {
+    const authData = req.user
+    if(authData.id !== +id){
+      throw new BadRequestException("yu cant update others' profile")
+    }
     return this.usersService.update(+id, updateUserDto);
   }
 
+  @UseGuards(AuthGuard)
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  remove(@Param('id') id: string,@Request() req:any) {
+    const authData = req.user
+    if(authData.id !== +id){
+      throw new BadRequestException("yu cant update others' profile")
+    }
     return this.usersService.remove(+id);
   }
 }
